move 'start_new_session' to channel
